+++ src/cpu/VirtualFluidsCore/BoundaryConditions/BCArray3D.cpp
@@ -173,11 +173,11 @@ std::string BCArray3D::toString() const
    std::size_t bcCounter = 0;
    std::size_t undefCounter = 0;
 
-   for (int x1 = 0; x1 < bcindexmatrix.getNX1(); x1++)
+   for (size_t x1 = 0; x1 < bcindexmatrix.getNX1(); x1++)
    {
-      for (int x2 = 0; x2 < bcindexmatrix.getNX2(); x2++)
+      for (size_t x2 = 0; x2 < bcindexmatrix.getNX2(); x2++)
       {
-         for (int x3 = 0; x3 < bcindexmatrix.getNX3(); x3++)
+         for (size_t x3 = 0; x3 < bcindexmatrix.getNX3(); x3++)
          {
             if (bcindexmatrix(x1, x2, x3) >= 0) bcCounter++;
             else if (bcindexmatrix(x1, x2, x3) == FLUID) fluidCounter++;
